Change how the Custom mirror file is handled
Save the Custom file in /var/lib/pacman-mirrors, instead of /etc/pacman.d/mirrors. This way we can assume that the default mirrors dir is clean.
If we need to comment the Custom country in the configuration file, do it after generating the mirrorlist.
No need to remove the Custom file anymore.